refactor: 라우팅 설정 파일 단위 분리 및 구조 개선
