+++ src/routes/router.tsx
@@ -1,43 +1,43 @@
 import { Navigate, Route, Routes } from 'react-router';
-import { MainPage } from './main/page';
-import { KakaoRegisterPage } from './register/page';
-import KakaoRedirect from './register/auth/callback/page';
-import { HumanRightsPage } from './human-rights/page';
-import { HumanRightsDetailPage } from './human-rights/[id]/page';
-import { HumanRightsEditPage } from './human-rights/edit/page';
-import { SuggestPage } from './sug-notice/page';
-import { SuggestDetailPage } from './sug-notice/[id]/page';
-import { SuggestWritePage } from './sug-notice/edit/page';
-import { NoticePage } from './notice/page';
-import { NoticeEditPage } from './notice/edit/page';
-import NoticeDetailPage from './notice/[id]/page';
-import NoticePatchPage from './notice/patch/page';
-import CampusPage from './campus/page';
-import PersonalDataPage from './personal-data/page';
-import { MainLayout } from './layout/MainLayout';
-import { KakaoRegisterRedirectPage } from './register/redirect/page';
-import MyPage from './mypage/page';
-import DataPage from './data/page';
-import DataDetailPage from './data/[id]/page';
-import DataEditPage from './data/edit/page';
-import ProfilePage from './mypage/profile/page';
-import MyPostsPage from './mypage/myPosts/page';
-import { ServiceNoticePage } from './mypage/service-notice/page';
-import { ServiceNoticeDetailPage } from './mypage/service-notice/[id]/page';
-import { ServiceNoticeEditPage } from './mypage/service-notice-edit/page';
-import ServiceNoticePatchPage from './mypage/service-notice/service-noticePatch/page';
-import { QnApage } from './qna-notice/page';
-import QnaDetailPage from './qna-notice/[id]/page';
-import QnaEditPage from './qna-notice/edit/page';
-import { TOSPage } from './register/step/tos/page';
-import { OnboardingPage } from './register/step/onboarding/page';
-import { CertifyErrorPage } from './register/step/errorcheck/page';
-import { CertifyApplyPage } from './register/step/errorapply/page';
-import { GeneralLoginPage } from './register/step/scouncil/page';
-import { RegisterLayout } from './layout/registerLayout';
-import IntroPage from './intro/page';
-import { SchedulePage } from './schedule/page';
-import { ScheduleEditPage } from './schedule/edit/page';
+import { MainPage } from '../pages/main/page';
+import { KakaoRegisterPage } from '../pages/register/page';
+import KakaoRedirect from '../pages/register/auth/callback/page';
+import { HumanRightsPage } from '../pages/human-rights/page';
+import { HumanRightsDetailPage } from '../pages/human-rights/[id]/page';
+import { HumanRightsEditPage } from '../pages/human-rights/edit/page';
+import { SuggestPage } from '../pages/sug-notice/page';
+import { SuggestDetailPage } from '../pages/sug-notice/[id]/page';
+import { SuggestWritePage } from '../pages/sug-notice/edit/page';
+import { NoticePage } from '../pages/notice/page';
+import { NoticeEditPage } from '../pages/notice/edit/page';
+import NoticeDetailPage from '../pages/notice/[id]/page';
+import NoticePatchPage from '../pages/notice/patch/page';
+import CampusPage from '../pages/campus/page';
+import PersonalDataPage from '../pages/personal-data/page';
+import { MainLayout } from '../pages/layout/MainLayout';
+import { KakaoRegisterRedirectPage } from '../pages/register/redirect/page';
+import MyPage from '../pages/mypage/page';
+import DataPage from '../pages/data/page';
+import DataDetailPage from '../pages/data/[id]/page';
+import DataEditPage from '../pages/data/edit/page';
+import ProfilePage from '../pages/mypage/profile/page';
+import MyPostsPage from '../pages/mypage/myPosts/page';
+import { ServiceNoticePage } from '../pages/mypage/service-notice/page';
+import { ServiceNoticeDetailPage } from '../pages/mypage/service-notice/[id]/page';
+import { ServiceNoticeEditPage } from '../pages/mypage/service-notice-edit/page';
+import ServiceNoticePatchPage from '../pages/mypage/service-notice/service-noticePatch/page';
+import { QnApage } from '../pages/qna-notice/page';
+import QnaDetailPage from '../pages/qna-notice/[id]/page';
+import QnaEditPage from '../pages/qna-notice/edit/page';
+import { TOSPage } from '../pages/register/step/tos/page';
+import { OnboardingPage } from '../pages/register/step/onboarding/page';
+import { CertifyErrorPage } from '../pages/register/step/errorcheck/page';
+import { CertifyApplyPage } from '../pages/register/step/errorapply/page';
+import { GeneralLoginPage } from '../pages/register/step/scouncil/page';
+import { RegisterLayout } from '../pages/layout/RegisterLayout';
+import IntroPage from '../pages/intro/page';
+import { SchedulePage } from '../pages/schedule/page';
+import { ScheduleEditPage } from '../pages/schedule/edit/page';
 
 export function MainRouter() {
   return (
